feat(robin): Column conversion, F.* stubs, agg aliases; remove test skips; analysis
- Column conversion: py_any_to_select_expr/column accept obj.name (Sparkless Column)
- _robin_sql: _to_robin_select_item() for select/filter/withColumn; Sparkless Column -> F.col(name)
- F.*: stubs for rlike/with_field when crate missing; add stubs for approx_count_distinct, stddev,
  array_contains, posexplode_outer, date_trunc, input_file_name, size, element_at, nvl, etc.
- groupBy().avg/sum/min/max/count: alias aggregation columns (avg(col), sum(col), ...) for PySpark parity
- conftest: remove Robin skip list and PySpark skip in fixtures; keep archive skips
- tests: add TEST_RUN_ANALYSIS.md
- Cargo: robin-sparkless 0.17, create_dataframe_from_rows(,,false), write_delta(,,false), round scale i32

+++ src/pyfunctions.rs
@@ -3,8 +3,8 @@
 use pyo3::exceptions::PyValueError;
 use pyo3::prelude::*;
 use pyo3::types::PyList;
-use robin_sparkless_polars::column::Column as RobinColumn;
-use robin_sparkless_polars::functions;
+use robin_sparkless::column::Column as RobinColumn;
+use robin_sparkless::functions;
 
 use crate::pycolumn::{py_any_to_column, py_any_to_select_expr, PyColumn};
 use crate::pysortorder::PySortOrder;
@@ -211,7 +211,7 @@ pub fn floor(col: &Bound<'_, PyAny>) -> PyResult<PyColumn> {
 #[pyo3(signature = (col, decimals=0))]
 pub fn round(col: &Bound<'_, PyAny>, decimals: u32) -> PyResult<PyColumn> {
     let c = py_any_to_select_expr(col)?;
-    Ok(PyColumn::from_robin(functions::round(&c, decimals)))
+    Ok(PyColumn::from_robin(functions::round(&c, decimals as i32)))
 }
 
 /// F.sqrt(col)
